refactor gen-pr to separate github related functions from workflow
